separates out different checklists on registration form into own components
